Consolidated all junit required event listening
